backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
celery[redis]>=5.4,<6.0
requests>=2.31,<3.0

numpy>=1.26,<3.0
# opencv-python>=4.10,<5.0
mediapipe>=0.10.14,<0.11
PyMuPDF>=1.24,<2.0
Pillow>=10.4,<12.0
